Create maven.yml
Adds a GitHub action so all PRs (and direct pushes to main) require passing tests in order to be merged with main. I'm not 100%, but we might need to introduce a new step for integration testing and beyond.

I don't have admin permissions, so in order for this to apply, it will need to be set as a branch rule.